TEAM2-16 제미나이 1차 리뷰 반영 및 useMapDrawing 훅, imageUpload 타입 제작
