+++ src/components/MapArea.tsx
@@ -1,8 +1,9 @@
-import React, { useState } from "react";
+import React from "react";
 import { Circle, Pentagon } from "lucide-react";
 import Button from "./Button";
 import { Label } from "./label";
-import type { AreaData, PolygonPoint, CircleData } from "../types/mapArea";
+import type { AreaData } from "../types/mapArea";
+import { useMapDrawing } from "../hooks/useMapDrawing";
 
 import "../styles/MapArea.css";
 
@@ -12,96 +13,23 @@ interface MapAreaProps {
 }
 
 const MapArea: React.FC<MapAreaProps> = ({ className, onAreaChange }) => {
-  // Area drawing states
-  const [areaType, setAreaType] = useState<"circle" | "polygon" | null>(null);
-  const [polygonPoints, setPolygonPoints] = useState<PolygonPoint[]>([]);
-  const [circleData, setCircleData] = useState<CircleData | null>(null);
-  const [isDrawing, setIsDrawing] = useState(false);
-  const [mapType, setMapType] = useState<"map" | "satellite">("map");
-  const [mousePosition, setMousePosition] = useState<{
-    x: number;
-    y: number;
-  } | null>(null);
-
-  const handleAreaTypeSelect = (type: "circle" | "polygon") => {
-    // 이미 선택된 타입을 다시 클릭하면 취소
-    if (areaType === type) {
-      setAreaType(null);
-      setPolygonPoints([]);
-      setCircleData(null);
-      setIsDrawing(false);
-      onAreaChange?.(null);
-    } else {
-      setAreaType(type);
-      setPolygonPoints([]);
-      setCircleData(null);
-      setIsDrawing(true);
-    }
-  };
-
-  const handleMapClick = (e: React.MouseEvent<HTMLDivElement>) => {
-    if (!isDrawing || !areaType) return;
-
-    const rect = e.currentTarget.getBoundingClientRect();
-    const x = e.clientX - rect.left;
-    const y = e.clientY - rect.top;
-
-    if (areaType === "polygon") {
-      const newPoints = [...polygonPoints, { x, y }];
-      setPolygonPoints(newPoints);
-      onAreaChange?.({ type: "polygon", points: newPoints });
-    } else if (areaType === "circle") {
-      if (!circleData) {
-        setCircleData({ center: { x, y }, radius: 0 });
-      } else {
-        const radius = Math.sqrt(
-          Math.pow(x - circleData.center.x, 2) +
-            Math.pow(y - circleData.center.y, 2),
-        );
-        const newCircleData = { ...circleData, radius };
-        setCircleData(newCircleData);
-        setIsDrawing(false);
-        onAreaChange?.({ type: "circle", data: newCircleData });
-      }
-    }
-  };
-
-  const handleMapMouseMove = (e: React.MouseEvent<HTMLDivElement>) => {
-    if (!isDrawing || !areaType) return;
-
-    const rect = e.currentTarget.getBoundingClientRect();
-    const x = e.clientX - rect.left;
-    const y = e.clientY - rect.top;
-
-    setMousePosition({ x, y });
-  };
-
-  const handleMapMouseLeave = () => {
-    setMousePosition(null);
-  };
-
-  const finishPolygon = () => {
-    if (polygonPoints.length >= 3) {
-      setIsDrawing(false);
-      onAreaChange?.({ type: "polygon", points: polygonPoints });
-    }
-  };
-
-  const clearArea = () => {
-    setPolygonPoints([]);
-    setCircleData(null);
-    setIsDrawing(false);
-    setAreaType(null);
-    onAreaChange?.(null);
-  };
-
-  const handleMapTypeChange = (type: "map" | "satellite") => {
-    setMapType(type);
-  };
-
-  const hasAreaData = polygonPoints.length > 0 || circleData;
-  const canFinishPolygon =
-    areaType === "polygon" && polygonPoints.length >= 3 && isDrawing;
+  const {
+    areaType,
+    polygonPoints,
+    circleData,
+    isDrawing,
+    mapType,
+    mousePosition,
+    hasAreaData,
+    canFinishPolygon,
+    handleAreaTypeSelect,
+    handleMapClick,
+    handleMapMouseMove,
+    handleMapMouseLeave,
+    finishPolygon,
+    clearArea,
+    handleMapTypeChange,
+  } = useMapDrawing({ onAreaChange });
 
   return (
     <div className={`map-area ${className || ""}`}>
@@ -140,6 +68,7 @@ const MapArea: React.FC<MapAreaProps> = ({ className, onAreaChange }) => {
           type="button"
           onClick={clearArea}
           className={`map-area-control-btn ${hasAreaData ? "active" : "disabled"}`}
+          disabled={!hasAreaData}
         >
           영역 지우기
         </Button>
